% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_it}
\alias{calc_it}
\title{calc_it}
\usage{
calc_it(seeds, poly, poly_, poly_roots, c_calculation_depth, digits)
}
\arguments{
\item{seeds}{complex vector}

\item{poly}{numeric vector  # polynom coefficients}

\item{poly_}{numeric vector # derivative coefficients of the polynom => deriv(poly)}

\item{poly_roots}{complex roots vector # roots of the polynom => solve(poly)}

\item{c_calculation_depth}{#how many iterrations steps (depth of calculation)}

\item{digits}{how many decimal digit till calculation stop}
}
\description{
Berechntet die Funktionswerte eines Polynoms für gegebenen Vektor.
}
\examples{
polynom <-polynom::polynomial(coef = c(-0.2,-1,1,0,0,1))
roots <- solve(polynom)
seeds <- expand.grid(seq(-2,2, length.out = 200),seq(-2,2, length.out = 200))
seeds <- complex(real = seeds$Var1, imaginary = seeds$Var2)
calc_it(seeds, polynom, deriv(polynom), roots,100,6)
}
