% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_roots_from_seeds_C}
\alias{calc_roots_from_seeds_C}
\title{calc_roots_from_seeds_C}
\arguments{
\item{seeds}{seeds}

\item{roots}{roots from polynom equation}

\item{poly}{polynomial coefficients}
}
\description{
Create Student data frame.
}
\details{
calc_roots_from_seeds_C
}
\examples{
library(Rcpp)
df <- expand.grid(re = seq(0,10,0.1), im = seq(-10,0,0.1))
seeds <- complex(df$re,df$im)
calc_roots_from_seeds_C(seeds, c(0.2,2.5,3+1i), 1:10)
}
