% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calculate_cpk}
\alias{calculate_cpk}
\title{Process capability index}
\usage{
calculate_cpk(LSL, USL, mue, sigma)
}
\arguments{
\item{LSL}{lower specification limit parameter}

\item{USL}{upper specification limit parameter}

\item{mue}{the estimated mean of the process \eqn{\mu}}

\item{sigma}{standard deviation \eqn{\sigma}}
}
\value{
vector
}
\description{
Calculation of Cpk Process capability index given parameter USL, LSL mean \eqn{\mu} and standard deviation \eqn{\sigma}.
}
\details{
Estimates what the process is capable of producing, considering that the process mean may not be centered between the specification (Upper and Lower) limits.

\url{https://en.wikipedia.org/wiki/Process_capability_index}
}
\examples{
calculate_cpk(LSL = -0.03, USL = -0.01, mue = -0.02, sigma = 0.005)
calculate_cpk(LSL = -0.05, USL = 0.05, mue =c(-0.01, -0.015), sigma = c(0.001,0.015))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
