% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacas.R
\name{cas_intersection_plane_line}
\alias{cas_intersection_plane_line}
\title{CAS intersection of a plane and a line}
\usage{
cas_intersection_plane_line(p, n, s, w)
}
\arguments{
\item{p}{position vector for the plane}

\item{n}{normal vector of the plane}

\item{s}{position vector of the line}

\item{w}{direction vector of the line}
}
\value{
vector of intersection point
}
\description{
Function returning intersecting point of a plane and a 3D-line.
}
\details{
The calculation is done by Ryacas (yacas) and uses symbolic math. The package Ryacas and the software YACAS needs to be installed.

\url{http://www.yacas.org/}
}
\examples{
library(Ryacas)
cas_intersection_plane_line(p = c(2,3,5),n = c(1,-5,-8),s = c(1,1,1), w = c(1,1,1))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
