% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacas.R
\name{cas_plane_fun}
\alias{cas_plane_fun}
\title{CAS coordinate function of a plane}
\usage{
cas_plane_fun(p, n)
}
\arguments{
\item{p}{position vector}

\item{n}{normal vector of the plane}
}
\value{
Yacas function
}
\description{
Function returning a cas function of a plane defined by position vector \eqn{p} and the normal vector \eqn{n} of a plane.
}
\details{
The calculation is done by Ryacas (yacas) and uses symbolic math. The package Ryacas and the software YACAS needs to be installed.

\url{http://www.yacas.org/}
}
\examples{
library(Ryacas)
cas_plane_fun(c(2,3,4), c(12,5,1)|>math_unit_vector())
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
