% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas.R, R/yacas.R
\name{cas_rot_matrix2d}
\alias{cas_rot_matrix2d}
\title{2d rotation matrix}
\usage{
cas_rot_matrix2d(x)

cas_rot_matrix2d(x)
}
\arguments{
\item{x}{vector of an angel in radiant}
}
\value{
Returns rotation matrix for given angle

Returns rotation matrix for given angle
}
\description{
Calculates a 2D rotation matrix with a given angle \code{x}

Calculates a 2D rotation matrix with a given angle \code{x}
}
\details{
The calculation is done by Ryacas (yacas) and uses symbolic math. The package Ryacas and the software YACAS \url{http://www.yacas.org/}
needs to be installed.

The calculation is done by Ryacas (yacas) and uses symbolic math. The package Ryacas and the software YACAS \url{http://www.yacas.org/}
needs to be installed.
}
\examples{
library(Ryacas)
cas_rot_matrix2d(c(pi))
rot_matrix <- cas_rot_matrix2d(c(-pi)|>ysym())
rot_matrix
library(Ryacas)
cas_rot_matrix2d(c(pi))
rot_matrix <- cas_rot_matrix2d(c(-pi)|>ysym())
rot_matrix
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}

Florian Wagner
\email{florian.wagner@wagnius.ch}
}
