% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacas.R
\name{cas_rot_matrix3d}
\alias{cas_rot_matrix3d}
\title{CAS 3d rotation matrix}
\usage{
cas_rot_matrix3d(x, angle)
}
\arguments{
\item{x}{unit vector c(x1,x2,x3)|>ysym()}

\item{angle}{vector of angel in radiant c(angle)|>ysym()}
}
\value{
Returns 3D rotation matrix for given axis and angle.
}
\description{
Calculates a 3D rotation matrix from a given axis and angle. The axis is defined by \code{x|>ysym()}. The Rotation will be done around the axis defined by
\code{angle|>ysym()} and the root. The angle will be appied by the right hand rule.
}
\details{
The calculation is done by Ryacas (yacas) and uses symbolic math. The package Ryacas and the software YACAS \url{http://www.yacas.org/}
needs to be installed.
}
\examples{
library(Ryacas)
cas_rot_matrix3d(matrix(c(1,2,3)),c(-pi))
rot_matrix <- cas_rot_matrix3d(matrix(c(1,2,3))|>ysym(),c(-pi)|>ysym())
rot_matrix
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
