% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cufft}
\alias{cufft}
\title{Perform a 1D FFT using CUFFT}
\arguments{
\item{n}{Integer. The number of elements in the FFT.}

\item{inverse}{Logical. Whether to compute the inverse FFT (TRUE) or forward FFT (FALSE).}

\item{h_idata_re}{Numeric vector. The real part of the input data.}

\item{h_idata_im}{Numeric vector. The imaginary part of the input data.}

\item{h_odata_re}{Numeric vector. The real part of the output data.}

\item{h_odata_im}{Numeric vector. The imaginary part of the output data.}
}
\value{
A list with the real and imaginary parts of the transformed data.
}
\description{
This function uses CUDA's CUFFT library to compute the 1D FFT on a given set of data.
}
