% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{dict_assign_hash}
\alias{dict_assign_hash}
\title{Assigne key and value to fast dictionary}
\usage{
dict_assign_hash(
  x,
  value,
  pos = -1,
  envir = as.environment(pos),
  inherits = FALSE,
  immediate = TRUE
)
}
\value{
Environment hash that can be used as fast dictionary.
}
\description{
The reason for using dictionaries in the first place is performance.
Although it is correct that you can use named vectors and lists for the task,
the issue is that they are becoming quite slow and memory hungry with more data.
Yet what many people don't know is that R has indeed an inbuilt dictionary data structure
environments with the option hash = TRUE
}
\examples{
df <- data.frame(key   = c("ch","se","de","it"),
                 value = c(41L,46L,49L,39L))
# initialize hash
hash = dict_init_hash(nrow(df))
# Assign key and values
dict_assign_hash(df$key, df$value, hash)
dict_get_hash(c("ch", "it"), hash)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
