% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{dict_exists_key}
\alias{dict_exists_key}
\title{Check if key is in dictionary}
\usage{
dict_exists_key(
  x,
  where = -1,
  envir = if (missing(frame)) as.environment(where) else sys.frame(frame),
  frame,
  mode = "any",
  inherits = TRUE
)
}
\arguments{
\item{x}{key (character)}

\item{where}{= -1}

\item{envir}{= if (missing(frame)) as.environment(where) else sys.frame(frame)}

\item{frame}{frame}

\item{mode}{= "any"}

\item{inherits}{= TRUE}
}
\value{
Vector True/False
}
\description{
The reason for using dictionaries in the first place is performance.
Although it is correct that you can use named vectors and lists for the task,
the issue is that they are becoming quite slow and memory hungry with more data.
Yet what many people don't know is that R has indeed an inbuilt dictionary data structure
environments with the option hash = TRUE
}
\examples{
df <- data.frame(key   = c("ch","se","de","it"),
                 value = c(41L,46L,49L,39L))
hash <- dict_from_data.frame(df)
dict_exists_key(c("ch", "it", "xx"), hash)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
