% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{dict_from_data.frame}
\alias{dict_from_data.frame}
\title{Create a fast dictionary from data frame}
\usage{
dict_from_data.frame(df)
}
\arguments{
\item{df}{data frame with 1. column = key (must be character), 2. column = value (any atomic type)}
}
\value{
Environment hash that can be used as fast dictionary.
}
\description{
The reason for using dictionaries in the first place is performance.
Although it is correct that you can use named vectors and lists for the task,
the issue is that they are becoming quite slow and memory hungry with more data.
Yet what many people don't know is that R has indeed an inbuilt dictionary data structure
environments with the option hash = TRUE
}
\examples{
df <- data.frame(key   = c("ch","se","de","it"),
                 value = c(41L,46L,49L,39L))
hash <- dict_from_data.frame(df)
dict_get_values(c("ch", "it"), hash)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
