% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{dict_init}
\alias{dict_init}
\title{Initialize fast dictionary environment}
\usage{
dict_init(length)
}
\arguments{
\item{length}{length of the environment, e.g. nrow(data_frame)}
}
\value{
A initialized environment that can be assigned with keys and values
}
\description{
The reason for using dictionaries in the first place is performance.
Although it is correct that you can use named vectors and lists for the task,
the issue is that they are becoming quite slow and memory hungry with more data.
Yet what many people don't know is that R has indeed an inbuilt dictionary data structure
environments with the option hash = TRUE
}
\details{
\url{https://blog.ephorie.de/hash-me-if-you-can}
}
\examples{
df <- data.frame(key   = c("ch","se","de","it"),
                 value = c(41L,46L,49L,39L))
# initialize hash
hash = dict_init(nrow(df))
dict_assign_key_values(df$key, df$value, hash)
dict_get_values(c("ch", "it"), hash)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
