% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{dict_update}
\alias{dict_update}
\title{Update key/value pairs of fast dictionary}
\usage{
dict_update(df, dict)
}
\arguments{
\item{df}{data frame with column 1 = key and column 2 = value}

\item{dict}{hash}
}
\value{
Environment hash that can be used as fast dictionary.
}
\description{
The reason for using dictionaries in the first place is performance.
Although it is correct that you can use named vectors and lists for the task,
the issue is that they are becoming quite slow and memory hungry with more data.
Yet what many people don't know is that R has indeed an inbuilt dictionary data structure
environments with the option hash = TRUE
}
\details{
\url{https://blog.ephorie.de/hash-me-if-you-can}
}
\examples{
df <- data.frame(key   = c("ch","se","de","it"),
                 value = c(41L,46L,49L,39L))
hash <- dict_from_data.frame(df)
dict_get_values(c("ch", "it"), hash)
hash <- dict_update(data.frame(key = c("ch","it","test"), value = c(4.1, 0.39,0)), hash)
dict_get_values(ls(hash), hash)

}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
