% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expandGridComplex}
\alias{expandGridComplex}
\title{expandGridComplex}
\arguments{
\item{x}{complex vector}

\item{epsilon}{consider the value as zero}
}
\description{
Create Student data frame with all possible varians of a complex vector.
}
\details{
expandGridComplex
}
\examples{
expandGridComplex(complex(real = rnorm(100), imag = rnorm(100)))
}
