% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f_ableitung}
\alias{f_ableitung}
\title{f_ableitung}
\usage{
f_ableitung(x)
}
\arguments{
\item{x}{vector}
}
\description{
Berechntet die Funktionswerte von f(x) = y = -1 + 2\emph{x + 5}x^4.
}
\details{
f_ableitung
}
\examples{
f_ableitung(1+2i)
}
