% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f_vector}
\alias{f_vector}
\title{f_vector}
\usage{
f_vector(x)
}
\arguments{
\item{x}{vector}
}
\description{
Berechntet die Funktionswerte eines Polynoms.
}
\details{
f_vector
}
\examples{
f_vector(c(1+2i,-0.2-0.8i))
}
