% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_edges_2D}
\alias{find_edges_2D}
\title{find_edges_2D}
\arguments{
\item{c_search}{Logical vector}

\item{n_col}{x direction of the matrix}
}
\value{
logical vector if edge was found
}
\description{
Finds edges in an evenly spaced 2D grid.
}
\details{
find_edges_2D

Checks all neighbour indices in a vector \code{c_search} for \code{TRUE}. If all neighbours are \code{TRUE} it is not an edge, else it is.
}
\examples{
n_col <- 7
n_row <- 4
c_search <- c(
  0,1,0,0,0,1,0,
  1,1,1,1,1,1,1,
  0,1,0,1,1,1,1,
  0,0,1,1,1,1,1
  )
m_grid <- expand.grid(x = 0:(n_col-1), y = 0:(n_row-1))|>
  as.matrix()|>
  cbind(c_search)
m_grid|>
head()
m_grid|>
plot(col = m_grid[,"c_search"],
  xlim=c(0,n_col-1), ylim=c(0,n_row-1),
  pch = 19,
  cex = 2.5,
  main = "Input: c_search")

m_grid <- cbind(m_grid,
                edge = find_edges_2D(m_grid[,"c_search"], n_col))

m_grid|>
  head()
m_grid|>
  plot(col = m_grid[,"edge"],
       xlim=c(0,n_col-1), ylim=c(0,n_row-1),
       pch = 19,
       cex = 2.5,
       main = "Edge")
}
