% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_roots}
\alias{find_roots}
\title{find_roots}
\usage{
find_roots(x_n,roots,digits)
}
\arguments{
\item{x_n}{vector}

\item{roots}{vector}

\item{digits}{number of significant digits till calculation stops.}
}
\description{
finds roots in complex plane.
}
\details{
find roots
}
\examples{
#######
# seeds
seeds <- expand.grid(seq(-1,1, length.out = 20),
                      seq(-1,1, length.out = 20))
seeds <- complex(real = seeds$Var1, imaginary = seeds$Var2)

f <- function(x) 1 - x + x^2 + x^5
f_ableitung <- function(x) -1 + 2*x + 5*x^4

f1 <- polynom::polynomial(coef = c(-0.2,-1,1,0,0,1))

#######
x_n <- seeds|>
  lapply(function(x){
  x_n <- 1:100
  x_n[1] <- x
  for( i in 2:length(x_n)) {
    x_n[i] <- x_n[i-1]-(f(x_n[i-1])/f_ableitung(x_n[i-1]))
  }
return(x_n)
})
x_n[[1]]|>find_roots(summary(f1)$zeros,6)
}
