% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Flugbahn}
\alias{Flugbahn}
\title{Flugbahn}
\usage{
Flugbahn(v0,t,angle, target_vector, m, k,nb_return_values)
}
\arguments{
\item{v0}{Geschosss-Geschwindigkeit Vektor}

\item{t}{Zeit (interval der Berechnung)}

\item{angle}{Schusswinkel}

\item{target_vector}{target_vectorvektor}

\item{m}{Masse des Geschosses}

\item{k}{k-Wert = 0.5\emph{cw}rho*A: cw ==> Strömungswiderstandskoeffizient, Widerstandsbeiwert, rho ==> Dichte (kg/m^3), A ==> Fläche (m^2);}

\item{nb_return_values}{Anzahl Werte die zurückgegeben werden. Wenn Null gewählt wird dann werden alle brechneten Werte zurückgegenen.}
}
\value{
matrix

matrix t,v_x, v_y, s_x, s_y, distance (to target)
}
\description{
Berechnet die Flugbahn eines Geschosses mit berücksichtigung des Luftwiderstandes.
Es wird über den Zeitintervall "t" dv, und dx,dy berechnet.
}
\examples{
# target_vector
target_vector <- c(100, 20) # m
angle <- atan(target_vector[2]/target_vector[1]) #rad
# Zeitintervall
t <- 1e-5
# Geschoss
v0 = 55 # [m/s]
m <- 0.440 # [kg]
cw <- 0.18; # cw Wert des Geschosses (Nach Tabellen der Geschossform)
rho <- 1.293 # Dichte des Mediums (Luft) [kg/m^3]
A <- (0.22)^2*pi/4 #Projektions Flaeche in Flugrichtung [m^2]
k <- 0.5*cw*rho*A
Flugbahn_c <- Flugbahn(v0,t,angle,target_vector,m,k,1000)[,4:5]
colnames(Flugbahn_c) <- c("sx","sy")
Flugbahn_c|>
  plot(type = "l")
#tidyvers
library(tidyverse)
Flugbahn_c|>
  as_tibble()|>
  ggplot(aes(sx,sy))+
  geom_line()+
  geom_point(data = tibble(x = target_vector[1],y = target_vector[2]),aes(x,y), color = "red")+
  geom_text(aes(x = target_vector[1],y = target_vector[2]*1.1, label = "target_vector"))+
  coord_fixed()+
  labs(title = "Flugbahn Rcpp", subtitle = paste("v0:",v0,"\nt:",t))
}
