% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{function_calc_complex}
\alias{function_calc_complex}
\title{function_calc_complex}
\usage{
function_calc_complex(x,poly)
}
\arguments{
\item{x}{vector}

\item{poly}{polynom vector}
}
\description{
Berechntet die Funktionswerte eines Polynoms.
}
\examples{
function_calc_complex(1+2i,
polynom::polynomial(coef = c(-0.2, -1.0, 1.0, 0.0, 0.0, 1.0)))|>plot()
}
