% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{geo_convert_plane_coord_to_param}
\alias{geo_convert_plane_coord_to_param}
\title{Convert plane in coordinate form to parameter from}
\usage{
geo_convert_plane_coord_to_param(E, axis = c("x", "y", "z"))
}
\arguments{
\item{E}{Plane equation (Ryacas)}

\item{axis}{character vector indentifiying the axis variables  e.g. c("x1","x2","x3")}
}
\value{
list with matrix containing the Vectors
\deqn{\vec p, \vec u, \vec v}
and \deqn{\vec n = \begin{pmatrix} a \\ b \\ c \end{pmatrix},-d}
}
\description{
Get parametric function for a plane with the parameter s,t: \deqn{\vec x = \vec p + s \cdot \vec u + t \cdot \vec v}
The function returns a position vector \deqn{\vec p}, and two directions vectors \deqn{\vec u, \vec v}.
The Vector \deqn{\vec p} is also the normal vector (perpendicular to the plane).
}
\details{
The coordinate form of the plane is:
\deqn{a x + b y + c z - d = 0}
where \deqn{d = \vec p \cdot \vec n} and \deqn{\vec n = \begin{pmatrix} a \\ b \\ c \end{pmatrix}}
\url{https://de.wikipedia.org/wiki/Ebenengleichung#Normalenform}
\url{https://en.wikipedia.org/wiki/Euclidean_planes_in_three-dimensional_space}
}
\examples{
geo_convert_plane_coord_to_param("10*x+15*y+20*z+50")
geo_convert_plane_coord_to_param("-10*x+15*y+20*z+50")
geo_convert_plane_coord_to_param("10*x-15*y+20*z+50")
geo_convert_plane_coord_to_param("10*x+15*y-20*z+50")
geo_convert_plane_coord_to_param("10*x+15*y+20*z-50")
geo_convert_plane_coord_to_param("-10*x-15*y+20*z+50")
geo_convert_plane_coord_to_param("10*x-15*y-20*z+50")
geo_convert_plane_coord_to_param("-10*x+15*y-20*z+50")
geo_convert_plane_coord_to_param("-10*x-15*y+20*z-50")
geo_convert_plane_coord_to_param("10*x-15*y-20*z-50")
geo_convert_plane_coord_to_param("-10*x+15*y-20*z-50")
geo_convert_plane_coord_to_param("-10*x-15*y-50")
geo_convert_plane_coord_to_param("-10*z-15*y-50")
geo_convert_plane_coord_to_param("-10*z-15*x-50")
geo_convert_plane_coord_to_param("-10*x-15*y+50")
geo_convert_plane_coord_to_param("-10*z-15*y+50")
geo_convert_plane_coord_to_param("-10*z-15*x+50")
geo_convert_plane_coord_to_param("-10*z-50")
geo_convert_plane_coord_to_param("-10*y-50")
geo_convert_plane_coord_to_param("-10*x-50")
geo_convert_plane_coord_to_param("-10*z+50")
geo_convert_plane_coord_to_param("-10*y+50")
geo_convert_plane_coord_to_param("-10*x+50")
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
