% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{geo_interSec_3spheres}
\alias{geo_interSec_3spheres}
\title{Calculate the an intersection point of three spears}
\usage{
geo_interSec_3spheres(param_matrix, initial_guess)
}
\arguments{
\item{param_matrix}{for each spear a vector with its coefficients is needed: c(x0,y0,z0,r) so the matrix will have rows and four columns}

\item{initial_guess}{A vector with the intial guess c(x = 0,y = 0,z = 1)}
}
\value{
vector with named coordinates
}
\description{
Using numerical solver to calculate a intersecting point upon an initial guess.
}
\details{
The spears will be calculated from the form (x - x0)^2 + (y - y0)^2 + (z - z0)^2 - r^2).
}
\examples{
library(nleqslv)
params <- matrix(c(0 , 0, 0, 20.728,
                   15, 0, 0, 31.304,
                   0 ,10, 0, 26.576), ncol = 4, byrow = TRUE)
colnames(params) <- c("x0","y0","z0","r")
params
geo_interSec_3spheres(params, c(0,0,1))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
