% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{geo_intersec_circl_line}
\alias{geo_intersec_circl_line}
\title{Intersection point(s) of line and a circle}
\usage{
geo_intersec_circl_line(R, o, m, b)
}
\arguments{
\item{R}{Radius of the circle}

\item{o}{Position vector of the circle}

\item{m}{solpe of linear function}

\item{b}{intersect of linear function}
}
\value{
list of intersecting point(s)
}
\description{
Intersection point(s) of line and a circle
}
\details{
Computes the intersecting point of a line and a circle. The line is a linear function such as  \eqn{y=mx+b}. The circle in the form \eqn{(x-o_1)^2+(y-o_1)^2 = R^2}.
}
\examples{
R <- 1; o <- c(1.2,0.9); R <- 1; m <- 1; b <- 0.8;
geo_intersec_circl_line(R,o,m,b)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
