% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{geo_intersec_plane_line}
\alias{geo_intersec_plane_line}
\title{Intersection of a plane and a line}
\usage{
geo_intersec_plane_line(p, n, s, w)
}
\arguments{
\item{p}{position vector for the plane}

\item{n}{normal vector of the plane}

\item{s}{position vector of the line}

\item{w}{direction vector of the line}
}
\value{
vector of intersection point
}
\description{
Function returning intersecting point of a plane and a 3D-line if it exists.
}
\examples{
geo_intersec_plane_line(p = c(2,3,5),n = c(1,-5,-8),s = c(1,1,1), w = c(5,10,10))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
