% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{geo_slerp}
\alias{geo_slerp}
\title{slerp  by 3 points and a given radius.}
\usage{
geo_slerp(R, x1, x2, cp, nb_points = 10)
}
\arguments{
\item{R}{radius}

\item{x1}{starting point vector}

\item{x2}{end point vector}

\item{cp}{common centre point vector}

\item{nb_points}{count of points to be generate by the function.}
}
\value{
Matrix with \code{nrow(nb_points)}
}
\description{
Radius interpolation by 3 points and a given radius.
The Radius interpolation will be calculated using the \code{cp} = common centre point and the vector \code{x1} and \code{x2}.
}
\details{
The slerp calculation is done by the following formula:
\deqn{Slerp(\vec x_1, \vec x_2,t) = \frac{sin((1 - t) \cdot \Omega)}{sin(\Omega)} \cdot \vec x_1   + \frac{sin(t \cdot \Omega)}{sin(\Omega)}  \cdot  \vec x_2}
\url{https://en.wikipedia.org/wiki/Slerp}
}
\examples{
p <- t(as.matrix(data.frame(x1 = c(-10,-5,50),
                            x2 = c(-20,10,2),
                            cp = c(10,-10,10))))

m <- geo_slerp(R  =10,
               x1 = p["x1",],
               x2 = p["x2",],
               cp = p["cp",],
               nb_points = 20)
m <- rbind(p,m)
#Plot 3D
library(rgl)
plot3d( m[,1], m[,2], m[,3], type = "p", lwd = 2, top = TRUE,
        #col = rainbow(nrow(m)),
        aspect = c(diff(c(min(m[,1]),max(m[,1]))),
                   diff(c(min(m[,2]),max(m[,2]))),
                   diff(c(min(m[,3]),max(m[,3])))
                   ))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
