% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketing.R
\name{get_NPS}
\alias{get_NPS}
\title{Get Net Promoter Score}
\usage{
get_NPS(x, rating, c_group)
}
\arguments{
\item{rating}{string with column name of data frame containing the vector of the ratings}

\item{c_group}{string with column name to group the data frame and calcultae NPS for each group}
}
\description{
Get Net Promoter Score of rating 1...10. NPS = Ratio Promoters - Ratio Detractors,
whereas the Ratio of Promoters is calculated by count of Promoters / count total
and the the Ratio of Detractors is calculated by count of Detractors / count total.
}
\examples{
library(tidyverse)
# Test data
df_data <- tibble(user   = c(letters[1:10],rep("a",5)),
                   Feedback = c(1:10,rep(10,5)),
                   Gruppe = c(rep("A",10),rep("B",5)))\%>\%
  mutate(user = factor(user))

get_NPS(df_data, "Feedback", "Gruppe")
get_NPS(df_data, "Feedback")
}
