% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_functions.R
\name{lf_fromPoints}
\alias{lf_fromPoints}
\title{linear function from 2 points}
\usage{
lf_fromPoints(x, ...)
}
\arguments{
\item{x}{matrix or data frame of points just first point}

\item{...}{second point}
}
\value{
matrix with slope and intercept
}
\description{
get linear fuction from 2 points x1 and x2
}
\examples{
lf_fromPoints(c(0,0),c(1,1))
Points <- rbind(c(-2,5),
           c( 3,3))
result <- lf_fromPoints(Points)
result
plot(Points, asp = 0, xlab = "x", ylab = "y")
abline(result[,2],result[,1])
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
