% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_functions.R
\name{lf_intersect}
\alias{lf_intersect}
\title{Intersection point from two linear functions}
\usage{
lf_intersect(x, ...)
}
\arguments{
\item{x}{vector, matrix or data frame with slope and intercept}

\item{...}{vector with slope and intercept}
}
\value{
vector
}
\description{
Calculate intersecting point(s) from two linear function \code{x} and linear function \code{y}.
}
\details{
Get the intersecting point of two linear Functions. If only \code{x} is supplied, \code{x} needs to be a matrix or data.frame containing both
linear functions. If \code{x} and or \code{y} have names you shall use the names (slope, intercept).
}
\examples{
x <- c(slope = 0.5, intercept = 1)
y <- c(slope = -1, intercept = 2)
lf_intersect(x,y)
rbind(x,y)|>lf_intersect()
matrix(c(x,y),ncol = 2, byrow = TRUE)|>lf_intersect()
matrix(c(0.2,-1,2,3),ncol = 2)|>lf_intersect()
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
