% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_functions.R
\name{lf_lf}
\alias{lf_lf}
\title{Linear function: \eqn{f(x) = y = mx+b}}
\usage{
lf_lf(x, parameter, ...)
}
\arguments{
\item{x}{vector}

\item{parameter}{Either vector, matrix or data.frame with parameter c(\code{m},\code{b}) or just the slope \code{m}. It is also possible to supply a matrix as a set of parameter where each row represents c(\code{m},\code{b}).}

\item{...}{intercept \code{b}}
}
\value{
vector or matrix
}
\description{
linear function with parameter slope \code{m} and intercept \code{b}.
}
\details{
The function calculates \code{y} for given vector \code{x} with the \code{parameter},(\code{...}) supplied.
}
\examples{
lf_lf(-5:5,0.8,3.265)
lf_lf(-5:5,c(1,3))
df <- data.frame(x = c(0.5,1,-0.5), y = c(1,0,-1))
lf_lf(-5:5,df)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
