% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_functions.R
\name{lf_perpendicular}
\alias{lf_perpendicular}
\title{Find linear function perpendicular to linear function through given point}
\usage{
lf_perpendicular(x, lf, ...)
}
\arguments{
\item{x}{point vector}

\item{lf}{vector with slope and intercept of linear function or just the slope if intercept will be supplied by \code{...}}

\item{...}{intercept}
}
\value{
matrix with slope and intercept
}
\description{
Find linear function`s perpendicular to supplied linear function through given point \code{x}.
}
\examples{
lf_perpendicular(x = c(2,-9),
                      data.frame(slope = c(-3),intercept = c(5)))
lf_perpendicular(x = c(2,-9),
                      cbind(-3,5))
lf_perpendicular(x = c(2,-9),-3,5)
lf_perpendicular(x = c(2,-9), -3,5)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
