% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_functions.R
\name{lf_rev_slope}
\alias{lf_rev_slope}
\title{find perpendicular slope to linear function}
\usage{
lf_rev_slope(slope)
}
\arguments{
\item{slope}{The slop or m of linear function f(x)= mx+b}
}
\value{
data.frame(x,y) \code{x}
}
\description{
find perpendicular slope to linear function
}
\examples{
lf_rev_slope(c(0.1,-0.89))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
