% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_angle_quadrant}
\alias{math_angle_quadrant}
\title{Find quadrant(s) of vector(s)}
\usage{
math_angle_quadrant(x)
}
\arguments{
\item{x}{data.frame or matrix containing the coordinates.}
}
\value{
matrix of angle(s)
}
\description{
Find the angle(s) of vector(s) using the quadrant information.
}
\examples{
math_angle_quadrant(c(1,1))/pi*180
x <- expand.grid(x = -1:1, y = -1:1)|>as.matrix()
cbind(x, result = math_angle_quadrant(x))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
