% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_angle_quadrant_vector}
\alias{math_angle_quadrant_vector}
\title{Find the angle of a vector}
\usage{
math_angle_quadrant_vector(x)
}
\arguments{
\item{x}{vector containing the coordinates. e.g. First x, second y.}
}
\value{
Returns Quadrant(s) of vector(s)
}
\description{
find the
}
\details{
Find the angle of a vector using the quadrant information.
}
\examples{
math_angle_quadrant_vector(c( 1, 1))/pi*180
math_angle_quadrant_vector(c(-1, 1))/pi*180
math_angle_quadrant_vector(c(-1,-1))/pi*180
math_angle_quadrant_vector(c( 1,-1))/pi*180
math_angle_quadrant_vector(c( 0, 1))/pi*180
math_angle_quadrant_vector(c( 0,-1))/pi*180
math_angle_quadrant_vector(c( 1, 0))/pi*180
math_angle_quadrant_vector(c(-1, 0))/pi*180
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
