% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_angle_vector}
\alias{math_angle_vector}
\title{Find the angle of a vector}
\usage{
math_angle_vector(x)
}
\arguments{
\item{x}{vector containing the coordinates. e.g. First x, second y.}
}
\value{
Returns Quadrant(s) of vector(s)
}
\description{
find the
}
\details{
Find the angle of a vector takeing into account the quadrant.
}
\examples{
gfhfgh
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
