% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_cartesian_polar3D}
\alias{math_cartesian_polar3D}
\title{Cartesian to polar coordinates 3D}
\usage{
math_cartesian_polar3D(x)
}
\arguments{
\item{x}{vector of cartesian coordinates}
}
\value{
named vector with r (radius) , theta (angle to z-axis), phi(angle to x-axis)
}
\description{
Polar to Cartesian coordinate system transformation
}
\details{
\url{https://de.wikipedia.org/wiki/Kugelkoordinaten#Umrechnungen}
}
\examples{
math_cartesian_polar3D(c(1,1,1))
m <- expand.grid(x = -1:1, y = -1:1, z = -1:1)|>as.matrix()
result <- apply(m, 1, math_cartesian_polar3D)|>t()
colnames(result)<-c("r","theta","phi")
cbind(m, result)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
