% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_cartesian_to_polar}
\alias{math_cartesian_to_polar}
\title{Cartesian to polar coordinates}
\usage{
math_cartesian_to_polar(x)
}
\arguments{
\item{x}{vector}
}
\value{
named vector with r (radius) , theta (angle to x-axis in x/y plane), phi(angle to z in x/z plane)
}
\description{
Polar to Cartesian coordinate system transformation
}
\examples{
math_cartesian_to_polar(c(1,1,1))
m <- expand.grid(x = -1:1, y = -1:1, z = -1:1)|>as.matrix()
apply(m, 1, math_cartesian_to_polar)|>t()
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
