% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_conic_section_from_5points}
\alias{math_conic_section_from_5points}
\title{Conic section from five point}
\usage{
math_conic_section_from_5points(section_points, nb = 100)
}
\arguments{
\item{section_points}{matrix of section points, one point per row.}

\item{nb}{number of points returned}
}
\value{
list containing: data frame of calculated points, center...
}
\description{
calculate conic section from five given points \code{section_points}
}
\details{
\eqn{Ax^2+Bxy+Cy^2+Dx+Ey+F=0}

\url{https://en.wikipedia.org/wiki/Conic_section}
}
\examples{
section_points <- matrix(c(-1,3,
                          2,4,
                          6,2,
                          8,-5,
                          1,-3),
                          nrow = 5, byrow = T)
l_list <- math_conic_section_from_5points(section_points)
l_list
l_list$df|>plot(asp = 1)
abline(a = l_list$lf[1,2], b = l_list$lf[1,1])
abline(a = l_list$lf[2,2], b = l_list$lf[2,1])
points(x = l_list$center[1],y = l_list$center[2])

section_points <- matrix(c(-2,1,
                           0,2,
                           2,3,
                           0,-2,
                           2,0),
                           nrow = 5, byrow = T)
math_conic_section_from_5points(section_points)
}
\author{
Florian Wagner \email{florian.wagner@wagnius.ch}
}
