% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_cross_product}
\alias{math_cross_product}
\title{Compute Vector Cross Product}
\usage{
math_cross_product(x, y)
}
\arguments{
\item{x}{numeric vector or matrix}

\item{y}{numeric vector or matrix}
}
\value{
A scalar or vector of length the number of columns of x and y.
}
\description{
Computes the cross (or: vector) product of vectors in 3 dimensions.
In case of matrices it takes the first dimension of length 3 and computes the cross product between corresponding columns or rows.
}
\details{
The cross product of two vectors is the perpendicular vector to the plane spanned by the given vectors \code{x} and \code{y}.

\url{https://en.wikipedia.org/wiki/Cross_product}
}
\examples{
math_dot_product(1:5, 1:5)  #=> 55
# Length of space diagonal in 3-dim- cube:
sqrt(math_cross_product(c(1,1,1), c(1,1,1)))  #=> 1.732051
}
\seealso{
<math_dot_product>
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
