% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_dot_product}
\alias{math_dot_product}
\title{Compute the dot product of two vectors}
\usage{
math_dot_product(x, y, d = NULL)
}
\arguments{
\item{x}{Vector or a Matrix of vectors}

\item{y}{Vector or a Matrix of vectors}

\item{d}{Dimension along which to calculate the dot product}
}
\value{
Vector with length of dth dimension
}
\description{
If \code{x} and \code{y} are matrices, calculate the dot-product along the first non-singleton dimension.
If the optional argument d is given, calculate the dot-product along this dimension.
}
\examples{
math_dot_product(c(1,0,0),c(0.1,5,0),d = NULL)
math_dot_product(c(0.1,5,0),c(1,0,0),d = NULL)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
