% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{math_findNeighbours_2D}
\alias{math_findNeighbours_2D}
\title{math_findNeighbours_2D}
\arguments{
\item{c_search}{flat search vector to creat the matrix(c(),n_row,byrow = T) row wise sorted}

\item{n_col}{x direction of the matrix}
}
\description{
checks if all neighbour indices are there to detect an edge in a matrix
}
\details{
math_findNeighbours_2D
}
\examples{
n_col <- 7
 c_flat <- c(
  0,1,0,0,0,1,0,
  1,1,1,1,1,1,1,
  0,1,0,1,1,1,1,
  0,0,1,1,1,1,1)
m <- c_flat|>
  matrix(ncol = n_col, byrow = T)
m
df_data <- expand.grid(x = 1:n_col,  y = 1:4)
df_data$search <- c_flat == T
df_data$edge <- math_findNeighbours_2D(c_flat,n_col)
df_data$search == math_findNeighbours_2D(c_flat,n_col)
plot(df_data$x, df_data$y, col = df_data$search, main = "Search")
plot(df_data$x, df_data$y, col = df_data$edge, main = "Edges")

}
