% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_inbetweenAngle}
\alias{math_inbetweenAngle}
\title{Angle between 2 vectors}
\usage{
math_inbetweenAngle(u, v)
}
\arguments{
\item{u}{fist vector}

\item{v}{second vector}
}
\value{
Returns the smallest Angle between 2 vectors (scalar) in radiant.
\code{x}
}
\description{
Get the angle between 2 vectors. Returns the smallest Angle between 2 vectors in radiant.
}
\examples{
2D
u <- c(-0,1)
v <- c( 1,0)
math_inbetweenAngle(u,v)

3D
u <- c(-12, 13,   -2.56)
v <- c(  3,  5, -100   )
math_inbetweenAngle(u,v)

}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
