% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_functions.R
\name{math_lf}
\alias{math_lf}
\title{linear function with parameter mx+b}
\usage{
math_lf(x, m, b)
}
\arguments{
\item{x}{vector}

\item{m}{slope}

\item{b}{intercept}
}
\value{
vector \code{x}
}
\description{
linear function with parameter slope and intercept.
}
\details{
\eqn{f(x)=mx+b}
}
\examples{
math_lf(-12:10,0.8,3.265)
plot(math_lf(-12:10,0.8,3.265))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
