% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_lf_df_mb}
\alias{math_lf_df_mb}
\title{linear function but with data frame instead of single parameters}
\usage{
math_lf_df_mb(x, df_mb)
}
\arguments{
\item{x}{vector}

\item{df_mb}{data frame with "slope(s)" and "intercept(s)"}
}
\value{
vector \code{y}
}
\description{
Calculate y for different linear functions.
The parameters slope and intercept shall be provided by data frame with columns "slope" and "intercept".
}
\details{
f(x)= y = m_{slopes}* x + b_{intercepts}
}
\examples{
df <- data.frame(slope = 0.5, intercept = 1)
math_lf_df_mb(0:10,df)
df <- data.frame(slope = c(0.5,1,-0.5), intercept = c(1,-10,0))
math_lf_df_mb(10,df)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
