% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear functions.R
\name{math_lf_perpendicular}
\alias{math_lf_perpendicular}
\title{Find linear function perpendicular to linear function through given point}
\usage{
math_lf_perpendicular(x, lf)
}
\arguments{
\item{x}{point vector}

\item{lf}{data.frame with the definiton of the given linear function containing the column slope and intercept}
}
\value{
data.frame with the slope and intercept
}
\description{
Find linear function`s perpendicular to supplied linear function through given point \code{x}.
}
\examples{
math_lf_perpendicular(x = c(2,-9),
                      data.frame(slope = -3,intercept = 5))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
