% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_magnitude}
\alias{math_magnitude}
\title{Magnitude of a vector}
\usage{
math_magnitude(x)
}
\arguments{
\item{x}{numerical vector}
}
\value{
numerical vector of length one
\code{x}
}
\description{
Get the magnitude of a vector of any length
}
\examples{
math_magnitude(c(1,1))
math_magnitude(c(1,1,1))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
