% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_nonlinear_vector}
\alias{math_nonlinear_vector}
\title{Generate non-linear vector}
\usage{
math_nonlinear_vector(c_start, c_end, n)
}
\arguments{
\item{c_start}{value of first value of returned vector}

\item{c_end}{value of last index of returned vector}

\item{n}{lenght of returned vector}
}
\value{
polynomial character string
}
\description{
Function returns a vector of length \code{n}. The first value of the returned vector is \code{c_start} and the last value is \code{c_end}.
The caluclation is done accordingly: \eqn{c_{end}=c_{start}a^{n−1}} => \eqn{a=\frac{c_{end}}{c_{start}}^{\frac{1}{n-1}}}
so the returned vector will be caluclated with: \eqn{c_{start}\cdot a^{(0:(n-1))}}
}
\examples{
math_nonlinear_vector(1000,100000,8)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
