% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_polar_to_cartesian}
\alias{math_polar_to_cartesian}
\title{Polar to cartesian coordinates}
\usage{
math_polar_to_cartesian(r, theta, phi)
}
\arguments{
\item{r}{radius (distance from origin)}

\item{theta}{frist angle}

\item{phi}{second angle}
}
\value{
vector
}
\description{
Polar to cartesion coordinate system transformation
}
\examples{
math_polar_to_cartesian(1,pi/3,-pi/12)
m <- expand.grid(r = 1, theta = -1:1,phi = -1:1)|>as.matrix()
apply(m, 1, function(x){
    math_polar_to_cartesian(r = x[1], theta = x[2], phi = x[2])
  })|>t()
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
