% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_poly_coef}
\alias{math_poly_coef}
\title{Polynomial coefficients}
\usage{
math_poly_coef(equation)
}
\arguments{
\item{equation}{polynomial equation}

\item{var}{variable of the \code{equation}}
}
\value{
vector with polynomial coefficients.
}
\description{
Calculates the polynomial coefficients from a polynomial \code{equation}. The variablel \code{var} can be chosen.
The order of the terms does not matter.
}
\examples{
"-20*x + 4*x^2 + 5*x^3 + x^4"|>math_poly_coef()
"4*x^2 -  20*x + 5*x^3 + x^4"|>math_poly_coef()
"4*x^8 -  5*x^3 + x^4-0.2"|>math_poly_coef()
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
