% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_polyCoef}
\alias{math_polyCoef}
\title{Polynomial coefficients}
\usage{
math_polyCoef(equation, variable = "x")
}
\arguments{
\item{equation}{polynomial equation string}

\item{variable}{of the \code{equation}}
}
\value{
vector with polynomial coefficients.
}
\description{
Calculates the polynomial coefficients from a polynomial \code{equation}. The variablel \code{var} can be chosen.
The order of the terms does not matter.
}
\examples{
"-20*x + 4*x^2 + 5*x^3 + x^4"|>math_polyCoef()
"4*x^2 -  20*x + 5*x^3 + x^4"|>math_polyCoef()
"4*x^8 -  5*x^3 + x^4-0.2"|>math_polyCoef()
"3*z^3"|>math_polyCoef("z")
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
