% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_polynom_from_roots}
\alias{math_polynom_from_roots}
\title{Compute polynomial from roots}
\usage{
math_polynom_from_roots(roots, round_digits = 9)
}
\arguments{
\item{roots}{list of roots}

\item{round_digits}{the number of digit used to round the polynomial coefficients}
}
\value{
polynom as character string
}
\description{
The polynomial will be calculated according to the roots supplied in the list.
If roots are complex it will also include the complex conjugated root.
}
\examples{
c(-5,0,2i)|>math_polynom_from_roots()
c(-5,0,2i,-2i)|>math_polynom_from_roots()
c(3,2+1i)|>math_polynom_from_roots()
c(3,2+1i,2-1i)|>math_polynom_from_roots()
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
