% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_quadrant}
\alias{math_quadrant}
\title{Find quadrant of matrix or data frame}
\usage{
math_quadrant(x)
}
\arguments{
\item{x}{vector or matrix containing the coordinates. First x, second y.}
}
\value{
Returns Quadrant(s) of vector(s)
}
\description{
find the
}
\details{
Find the quadrant(s) of vector(s). If it is on principle axis it returns the angel.
}
\examples{
m <- matrix(c(c(1,-1,-1, 1,0,1),
            c(1, 1,-1,-1,1,0)),
            ncol = 2, byrow = FALSE)
math_quadrant(m)
data.frame(x = c(1,-1),y = c(1,-1))|>math_quadrant()
math_quadrant(c(1,0))
math_quadrant(c(1,0,12))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
