% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_quadrant_vector}
\alias{math_quadrant_vector}
\title{Find quadrant of single vector}
\usage{
math_quadrant_vector(x)
}
\arguments{
\item{x}{vector containing the coordinates. e.g.first x, second y.}
}
\value{
Returns Quadrant of vector(s)
}
\description{
find the
}
\details{
Finds the quadrant of vecotr. If it is on principle axis it returns the angle.
}
\examples{
x <- expand.grid(-1:1,-1:1)
x
x|>apply(1,math_quadrant_vector)
math_quadrant_vector(c("12","sdf"))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
