% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_rot_matrix2d}
\alias{math_rot_matrix2d}
\title{CAS 2d rotation matrix}
\usage{
math_rot_matrix2d(x)
}
\arguments{
\item{x}{vector of an angel in radiant}
}
\value{
Returns rotation matrix for given angle
}
\description{
Calculates a 2D rotation matrix with a given angle \code{x}
}
\details{
\url{https://en.wikipedia.org/wiki/Rotation_matrix}
}
\examples{
math_rot_matrix2d(c(pi))
math_rot_matrix2d(c(-pi))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
