% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_rot_matrix3d}
\alias{math_rot_matrix3d}
\title{2d rotation matrix}
\usage{
math_rot_matrix3d(x, angle)
}
\arguments{
\item{x}{unit vector c(x1,x2,x3)}

\item{winkel}{vector of angel in rad}
}
\value{
Returns matrix for given angle
}
\description{
calculates a 3D rotation matrix from axis and angle
}
\details{
\url{https://en.wikipedia.org/wiki/Rotation_matrix}
}
\examples{
math_rot_matrix2d(c(pi))
math_rot_matrix2d(c(-pi))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
