% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_rot_matrix_basic}
\alias{math_rot_matrix_basic}
\title{Basic rotation matrix}
\usage{
math_rot_matrix_basic(x, ax)
}
\arguments{
\item{x}{angel in radiant}

\item{ax}{parameter to define which principle axis will be used to calculate the rotation matrix}
}
\value{
Returns 3D basic rotation matrix for given axis and angle.
}
\description{
Calculates a 3D rotation matrix from a given principle axis and a given angle. The axis is defined by \code{ax}. The Rotation will be done around the axis defined by
\code{ax} and the root.
}
\details{
\url{https://en.wikipedia.org/wiki/Rotation_matrix}
}
\examples{
math_rot_matrix_basic(pi, "x")
math_rot_matrix_basic(pi, "x2")
math_rot_matrix_basic(pi, 3)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
