% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_rot_transform}
\alias{math_rot_transform}
\title{Transform vector by a rotation matrix}
\usage{
math_rot_transform(x, rot_matrix)
}
\arguments{
\item{x}{vector or matrix containing the coordinates}

\item{rot_matrix}{2D or 3D rotation matrix}
}
\value{
Returns 3D rotation matrix for given axis and angle.
}
\description{
transforms a given vecotr by a given rotation matrix using matirx multiplication
}
\details{
\code{rot_matrix}%*%\code{x}
\url{https://en.wikipedia.org/wiki/Rotation_matrix}
}
\examples{
rot_matrix <- math_rot_matrix3d(c(1,1,1), pi)
math_rot_transform(c(1,1,1),rot_matrix)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
