% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pendulum_motion}
\alias{pendulum_motion}
\title{pendulum_motion}
\arguments{
\item{L}{L채nge des Pendels (m)}

\item{delta_t}{(s)}

\item{THETA_0}{Startwinkel (rad)}

\item{THETA_DOT_0}{Anfangsgeschwindigkeit (m/s^2)}

\item{mu}{Reibungkoeffizient}

\item{calculation_stop}{Abbruch wenn die Geschwindigkeit kleiner und der Winkel kleiner als calculation_stop}

\item{nb_return_values}{Anzahl Werte die zur체ckgegeben werden. Wenn Null gew채hlt wird dann werden alle brechneten Werte zur체ckgegenen.}
}
\value{
matrix
}
\description{
Motion of pendulum.
}
\examples{
t = 35 ;
L = 2; delta_t = 1e-5; THETA_0 = pi/3; THETA_DOT_0 = 3; mu = 0.1;
calculation_stop = 0.05; nb_return_values = 800 ;
pendulum_motion(L, delta_t, THETA_0, THETA_DOT_0, mu, calculation_stop, nb_return_values)|>
plot(type = "l", main = "RccpTest: Pendulum Motion",xlab = "t", ylab = "angle")
pendulum_motion(L, delta_t, THETA_0, THETA_DOT_0, mu, calculation_stop, nb_return_values)[,c(1,3)]|>
plot(type = "l", main = "RccpTest: Pendulum Motion",xlab = "t", ylab = "angular velosity")
}
