% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{r_is.defined}
\alias{r_is.defined}
\title{Check if variable is defined in R`s global environment.}
\usage{
r_is.defined(sym)
}
\arguments{
\item{sym}{variable or function name}
}
\value{
TRUE or FALSE
}
\description{
Function to Check if variable is defined in R`s global environment.
}
