% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{r_map_data}
\alias{r_map_data}
\title{r_map_data}
\arguments{
\item{l}{list with column vectors}
}
\description{
convert ".asc" list to dataframe
}
\details{
r_map_data
}
\examples{
data = list(1:5,10:6,11:15)
data
r_map_data(data)
}
