% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{r_signif}
\alias{r_signif}
\title{Format number to defined signify(x,digits = 3)}
\usage{
r_signif(x, significant_digits = 3)
}
\arguments{
\item{x}{numerical vector}

\item{significant_digits}{number of significant digits you prefer.}
}
\value{
character vector
}
\description{
creates the same result as the Base R function format(\code{x}, format = "g", digits = 3), it`s just a wrapper with my preferred parameters.
}
\examples{
library(stringi)
test <- c(25,signif(rnorm(10),3))
test|>  paste0("\n")|>writeLines()
test|>
  r_signif()|>
  paste0("\n")|>writeLines()
test|>
  format(format = "g", digits = 3)|>
  paste0("\n")|>writeLines()
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
