% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmarkdown_utiilities.R
\name{r_toc_for_Rmd}
\alias{r_toc_for_Rmd}
\title{Automatically creates a table of contents for a .Rmd file}
\usage{
r_toc_for_Rmd(c_Rmd, create_nb = TRUE, nb_front = FALSE)
}
\arguments{
\item{c_Rmd}{Rmd file string}

\item{create_nb}{boolean to enable numbering for each heading.}

\item{nb_front}{boolean to have the numbering in front of each heading.}
}
\value{
.Rmd file string
}
\description{
Scans documents for headings and creates a table of contents (hyper linked). The returned string can directly be written as .Rmd file.
}
\details{
The function argument is a string of a R markdown .Rmd file which can be loaded via \code{readLines("fileName.Rmd")}
}
\examples{
print(tbl_of_contents.Rmd)
c_rmd <- r_toc_for_Rmd(tbl_of_contents.Rmd)
c_rmd
c_rmd <- r_toc_for_Rmd(tbl_of_contents.Rmd, TRUE)
c_rmd
}
