% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmarkdown_utiilities.R
\name{r_tocForRmdFiles}
\alias{r_tocForRmdFiles}
\title{Automatically creates a table of contents for all found .Rmd files in actual working directory}
\usage{
r_tocForRmdFiles(
  render_type = c("html_document"),
  toc_heading_string = "Table of Contents",
  pagebreak_level = "non",
  create_nb = TRUE,
  nb_front = TRUE,
  set_first_heading_level = FALSE,
  delete_Rmd_file = TRUE
)
}
\arguments{
\item{render_type}{choose the render type e.g. c("html_document","word_document","pdf_document")}

\item{toc_heading_string}{title string for TOC, e.g "Inhaltsverzeichnis", default is "Table of contents".}

\item{pagebreak_level}{Automatically add page breaks before new heading. Set the level to insert page breaks to e.g. c("non","1","2","3","4","5","6")}

\item{create_nb}{boolean to enable numbering for each heading.}

\item{nb_front}{boolean to have the numbering in front of each heading.}

\item{set_first_heading_level}{set first found heading level to heading 1}

\item{delete_Rmd_file}{boolean to delete the created .Rmd files.}
}
\description{
Automatically creates a table of contents for all found .Rmd files in actual working directory
}
\examples{
write(tbl_of_contents.Rmd)
r_tocForRmdFiles()
print(getwd())
}
