% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{round5Rappen}
\alias{round5Rappen}
\title{spez. Round  for Swiss currency "CHF"}
\usage{
round5Rappen(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector
}
\description{
rounds for swiss currency. Between values of 0.075 to 0.1249 will be round to 0.1.
From values of 0.025 to 0.0749 will be round to 0.1.
}
\examples{
round5Rappen(c( 0.02499, 0.025, 0.0749, 0.075))
round5Rappen(c(10.02499,10.025,10.0749,10.075))
round5Rappen(c(10.52499,10.525,10.5749,10.575))
round5Rappen(c(10.72499,10.725,10.7749,10.775))
}
