% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singal.R
\name{signal_center}
\alias{signal_center}
\title{Calculate unit vector}
\usage{
signal_center(x)
}
\arguments{
\item{x}{vector}
}
\value{
centred vector \code{x}
}
\description{
centre data point around zero, that the range of the vector \code{x} is +/- the same value.
}
\examples{
rnorm(100,mean = -0.5)|>signal_center()
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
