% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{sph2cart}
\alias{sph2cart}
\title{Sperical to cartesian coordinates}
\usage{
sph2cart(tpr)
}
\arguments{
\item{tpr}{t \eqn{\theta}, p \eqn{\phi}, r radius as vector or matrix}
}
\value{
vector
}
\description{
Transform sperical to cartesian coordinates according to international physics convention:
\eqn{\theta} in range 0...pi and \eqn{\phi} in range 0...2*pi
}
\details{
\url{https://de.wikipedia.org/wiki/Kugelkoordinaten#Umrechnungen}
}
\examples{
sph2cart(c(1,0.5,0.1))
rbind(c(1,0.5,0.1),
      c(1,0.2,0.6))|>
      sph2cart()
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
