% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Student}
\alias{Student}
\title{Student}
\arguments{
\item{ID}{last ID needed to create}

\item{Name}{Name of student}

\item{Lastname}{Lastname of student}

\item{eMail}{eMail}

\item{phone}{Phone}
}
\description{
Create Student data frame.
}
\details{
Student
}
\examples{
library(Rcpp)
Student(12)
Student(12, "Florian", "Wagner")
}
