% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_of_contents.Rmd}
\alias{tbl_of_contents.Rmd}
\title{Data for table of contents for a .Rmd file}
\format{
An object of class \code{character} of length 41.
}
\usage{
tbl_of_contents.Rmd
}
\description{
Data for r_tbl_contents
}
\keyword{datasets}
