% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{test.Rmd}
\alias{test.Rmd}
\title{Automatically creates a table of contents for a .Rmd file}
\usage{
test.Rmd()
}
\description{
R markdown data
}
\details{
The function argument is a stirng of a R markdown .Rmd file which can be loaded via \code{readLines("fileName.Rmd")}
}
