% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{truth.table}
\alias{truth.table}
\title{Generate truth.table}
\usage{
truth.table(
  bit,
  exponent = 2,
  hex_style = FALSE,
  stringsAsFactors = TRUE,
  KEEP.OUT.ATTRS = TRUE
)
}
\arguments{
\item{bit}{integer number of bits}

\item{exponent}{typ of truth table}

\item{hex_style}{If exponent > 9 the number will be changed to letters.}

\item{stringsAsFactors}{logical specifying if character vectors are converted to factors.}

\item{KEEP.OUT.ATTRS}{a logical indicating the "out.attrs" attribute. Check out expand.grid() function for more detailed information.}
}
\value{
A data frame containing one row for each combination of the supplied factors(\code{bit}\code{s}) .
}
\description{
Create a truth table for the number of \code{bit}`s supplied. The Standard is a binary truth.table so the \code{exponent = 2}.
If  (\code{exponent} > 9) you can choose (\code{hex_style = TRUE}) to exchange any bigger number than 9 with Letters,see examples for more Information.
}
\examples{
truth.table(3)
truth.table(3, 16)|>head(20)
truth.table(3, 16, hex_style = TRUE)|>head(20)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
