% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vector_function_calc_complex}
\alias{vector_function_calc_complex}
\title{vector_function_calc_complex}
\usage{
vector_function_calc_complex(x,poly)
}
\arguments{
\item{x}{vector}

\item{poly}{polynom vector}
}
\description{
Berechntet die Funktionswerte eines Polynoms für gegebenen Vektor.
}
\examples{
vector_function_calc_complex(-6:6,
polynom::polynomial(coef = c(-0.2, -1.0, 1.0, 0.0, 0.0, 1.0)))|>plot()
}
