% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expandGridComplex}
\alias{expandGridComplex}
\title{expandGridComplex}
\arguments{
\item{x}{complex vector}

\item{epsilon}{consider the imaginary value as zero if smaller then \code{epsilon}}
}
\value{
returns a complex vector
}
\description{
Create Student data frame with all possible variants of a complex vector.
}
\details{
expandGridComplex
}
\examples{
expandGridComplex(complex(real = 1:100, imag = rnorm(100)))
}
