% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_edges_3D}
\alias{find_edges_3D}
\title{Find edges in an evenly spaced 3D grid.}
\arguments{
\item{c_search}{Logical vector}

\item{n}{dimensions of 3D array c(nx, ny, nz)}
}
\value{
logical vector if edge was found
}
\description{
Checks all neighbour indices in a vector \code{c_search} for \code{TRUE}. If all neighbours are \code{TRUE} it is not an edge, else it is.
}
\examples{
n <- c(6L,5L,4L)
df_input <- expand.grid(
  x = 1:n[1],
  y = 1:n[2],
  z = 1:n[3])
df_input$search <- TRUE
df_input$edge <- find_edges_3D(df_input$search, n)
df_input$color <- ifelse(df_input$edge, "blue", "black")

df_input
df_input|>
  rgl::plot3d(col = df_input$color, size = 15,aspect = "iso")

}
