% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{geo_interSec_sph_line}
\alias{geo_interSec_sph_line}
\title{Computes intersection point of a line and a sphere surface.}
\usage{
geo_interSec_sph_line(p, u, m, r)
}
\arguments{
\item{p}{position vector for the line  c(x = ,y = ,z = )}

\item{u}{direction vector for the line c(x = ,y = ,z = )}

\item{m}{position vector of the centre of the sphere c(x = ,y = ,z = )}

\item{r}{radius of the sphere}
}
\value{
list of intersecting point(s) (x,y,z) and parameter \code{t}. Parameter \code{t} is giving the distance from \code{p} to \code{m}
}
\description{
Computes intersection point of a line and a sphere surface.
}
\examples{
p <- c(x = 1, y = -2, z = 3)
u <- c(x = 1, y = 0, z = 1)
m <- c(1,-2,4)
r <- 1
geo_interSec_sph_line(p, u, m, r)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
