% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{geo_slerp}
\alias{geo_slerp}
\title{slerp  by 3 points and a given radius.}
\usage{
geo_slerp(R, x1, x2, cp, nb_points = 10)
}
\arguments{
\item{R}{radius}

\item{x1}{starting point vector}

\item{x2}{end point vector}

\item{cp}{common centre point vector}

\item{nb_points}{count of points to be generate by the function.}
}
\value{
Matrix with \code{nrow(nb_points)}
}
\description{
Radius interpolation by 3 points and a given radius.
The Radius interpolation will be calculated using the \code{cp} = common centre point and the vector \code{x1} as the starting point and the vector \code{x2} as the end point.
}
\details{
The slerp calculation is done by the following formula:
\deqn{Slerp(p_0,p_1,t) = \frac{sin((1 - t) \cdot \Omega)}{sin(\Omega)} \cdot p_0   + \frac{sin(t \cdot \Omega)}{sin(\Omega)}  \cdot p_1}
\url{https://en.wikipedia.org/wiki/Slerp}
}
\examples{
p <- data.frame(x1 = c(-10,-5,50),
                x2 = c(-20,10,2),
                cp = c(10,-10,10))|>
   as.matrix()|>
   t()

m <- geo_slerp(R  =10,
               x1 = p["x1",],
               x2 = p["x2",],
               cp = p["cp",],
               nb_points = 20)
print(m)

#Plot 3D
library(rgl)
plot3d( m, type = "p", lwd = 2, top = TRUE,
        col = rainbow(nrow(m)),
        aspect = "iso")
plot3d(p, add = TRUE, col = "black")
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
