% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketing.R
\name{get_NPS}
\alias{get_NPS}
\title{Get Net Promoter Score}
\usage{
get_NPS(x, rating, c_group, df_levels)
}
\arguments{
\item{rating}{string with column name of the rating.}

\item{c_group}{string with the column name to group the data frame.}

\item{df_levels}{data frame containing the rating range for each type (Promoters, Passives, Detractors).}
}
\description{
Calculates the Net Promoter Score.
}
\details{
NPS = Ratio Promoters - Ratio Detractors,
whereas the Ratio of Promoters is calculated by count of Promoters / count total
and the the Ratio of Detractors is calculated by count of Detractors / count total.
  If the \code{c_group} parameter is available the NPS score will be calculated for each group.
If the \code{df_levels} parameter is provided the level to find Promoters, Passives and Detractors can be defined by this definition.
}
\examples{
library(tidyverse)
# Test data
df_data <- tibble(user   = c(letters[1:10],rep("a",5)),
                   Feedback = c(1:10,rep(10,5)),
                   Gruppe = c(rep("A",10),rep("B",5)))\%>\%
  mutate(user = factor(user))

get_NPS(df_data, "Feedback", "Gruppe")
get_NPS(df_data, "Feedback")

df_levels <- tibble(Detractors= c(1,3),Passives= c(4,4), Promoters= c(5,5))
df_data <- tibble(user   = c(letters[1:10],rep("a",5)),
                  Feedback = c(1:5,rep(5,10)),
                  Gruppe = c(rep("A",10),rep("B",5)))
get_NPS(df_data, "Feedback", "Gruppe",df_levels)
get_NPS(df_data, "Feedback",df_levels = df_levels)
}
