% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_functions.R
\name{lf_intersect}
\alias{lf_intersect}
\title{Intersection point from two linear functions}
\usage{
lf_intersect(x, ...)
}
\arguments{
\item{x}{matrix or data frame (column names (slope,intercept))}

\item{...}{matrix or data frame (column names (slope,intercept))}
}
\value{
vector
}
\description{
Calculate intersecting point(s) from two linear function \code{x} and linear function \code{y}.
}
\details{
Get the intersecting point of two linear Functions. If only \code{x} is supplied, \code{x} needs to be a matrix or data.frame containing both
linear functions. If \code{x} and or \code{y} have names you shall use the names (slope, intercept).
}
\examples{
x <- data.frame(slope = 2.3, intercept = -2)
y <- data.frame(slope = 2.3, intercept = -2)
lf_intersect(x,y)
y <- y|>
  rbind(data.frame(slope = 1.25, intercept = 2))
lf_intersect(y)
names(y)<-c("slope","intercept")
lf_intersect(y)
lf_intersect(data.frame(slope = c(0.2,-1),intercept = c(2,3)))
lf_intersect(matrix(c(0.2,-1,2,3),ncol = 2))
y <- matrix(c(0.2,-1,2,3),ncol = 2)
lf_intersect(y)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
