% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_betrag}
\alias{math_betrag}
\title{Magnitude of a vector}
\usage{
math_betrag(x)
}
\arguments{
\item{x}{numerical vector}
}
\value{
numerical vector of length one
\code{x}
}
\description{
Get the magnitude of a vector of any length
}
\details{
Get the magnitude of a vector of any lenght
}
\examples{
math_betrag(c(1,1))
math_betrag(c(1,1,1))
}
\author{
Florian Wagner
}
