% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_betrag}
\alias{math_betrag}
\title{Magnitude of a vector}
\usage{
math_betrag(x)
}
\arguments{
\item{x}{numerical vector}
}
\value{
numerical vector of \code{\link[=length]{length()}} one
}
\description{
Get the magnitude of a vector of any length.

This function is the same as \code{\link[=math_magnitude]{math_magnitude()}}
}
\examples{
math_betrag(c(1,1))
math_betrag(c(1,1,1))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
