% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_circle_from3points}
\alias{math_circle_from3points}
\title{Circle from 3 points}
\usage{
math_circle_from3points(x, type = "tibble")
}
\arguments{
\item{x}{numerical point matrix}

\item{type}{c("tibble","vector")  to choose the return class}
}
\value{
tibble or vector
\code{x}
}
\description{
Get the circle using 3 points.The function returns a tibble (dataframe) or
a named vector with the center point and the radius.
}
\details{
get center and radius from point matrix
}
\examples{
Matrix
x <- matrix(c(c(-2.23 , 4.389 ),
              c( 1.001,-3.23  ),
              c(  15.5,-7.2365)),
            nrow = 3, byrow = TRUE)

math_circle_from3points(x)

math_circle_from3points(x, type = "vector")
}
\author{
Florian Wagner
}
