% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_dot_product}
\alias{math_dot_product}
\title{Compute the dot product of two vectors}
\usage{
math_dot_product(x, y)
}
\arguments{
\item{x}{Vector or a Matrix of vectors}

\item{y}{Vector or a Matrix of vectors}
}
\value{
Vector with length of input dimension
}
\description{
'dot' or 'scalar' product of vectors or pairwise columns of matrices.
}
\details{
Returns the 'dot' or 'scalar' product of vectors or columns of matrices. Two vectors must be of same length, two matrices must be of the same size.
If x and y are column or row vectors, their dot product will be computed as if they were simple vectors.
}
\examples{
math_dot_product(c(1,0,0),c(0.1,5,0))
math_dot_product(c(1,0),c(0,1))
sqrt(math_dot_product(c(1,1,1), c(1,1,1)))  #=> 1.732051
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
