% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_lf_fromPoints}
\alias{math_lf_fromPoints}
\title{linear function from 2 points}
\usage{
math_lf_fromPoints(lf)
}
\arguments{
\item{lf}{data.frame(x1,x2) or matrix[x1,x2]}
}
\value{
data.frame(slope, intercept) \code{x}
#' @author Florian Wagner
\email{florian.wagner@wagnius.ch}
}
\description{
get linear fuction from 2 points
}
\examples{
math_lm_from_Points(data.frame(c(-2,5), c(3,5)))
}
