% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_lf_fromPoints}
\alias{math_lf_fromPoints}
\title{linear function from 2 points}
\usage{
math_lf_fromPoints(x)
}
\arguments{
\item{x}{matrix or data frame}
}
\value{
data.frame(slope, intercept) \code{x}
}
\description{
get linear fuction from 2 points x1 and x2
}
\examples{
x <- matrix(c(c(-2,5),
              c( 3,3)), nrow = 2, byrow = FALSE)

y <- data.frame(p1 = c(x = -1, y = -5),
                p2 = c(x =  3, y =  3))

math_lf_fromPoints(x)

result <- math_lf_fromPoints(y)

library(tidyverse)
ggplot(as_tibble(t(y)),aes(x,y))+
geom_abline(data = result,aes(slope = slope, intercept =intercept), size = 1.5)+
geom_point(shape = 10, size = 10,color = "blue")+
geom_hline(yintercept = 0)+
geom_vline(xintercept = 0)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
