% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_lf_intersect}
\alias{math_lf_intersect}
\title{Intersection point from two linear functions}
\usage{
math_lf_intersect(x, y)
}
\arguments{
\item{x}{matrix or data frame with columns "slope" and "intersect"}

\item{y}{matrix or data frame with columns "slope" and "intersect"}
}
\value{
matrix with intersecting point
}
\description{
Calculate the intercection point from two linear functions.
}
\examples{
x <- data.frame(slope = 2.3, intercept = -2)
y <- data.frame(slope = 2.3, intercept = -2)
math_lf_intersect(x,y)
y <- data.frame(slope = 1.25, intercept = 2)
math_lf_intersect(x,y)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
