% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_lf_perpendicular}
\alias{math_lf_perpendicular}
\title{find linear function perpendicular to linear function}
\usage{
math_lf_perpendicular(point, lf)
}
\arguments{
\item{point}{vector}

\item{lf}{data.frame(slope = c(0.12,0.78),
intercept = c(-25, 13))}
}
\value{
data.frame(x,y) \code{x}
}
\description{
find linear function perpendicular to another linear function defined by a given point.
}
\examples{
math_lf_perpendicular(point = c(2,-9),
                      data.frame(slope =c(-2,5),intercept = c(3,5)))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
