% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_quadrant}
\alias{math_quadrant}
\title{Find quadrant(s) of vector(s)}
\usage{
math_quadrant(x)
}
\arguments{
\item{x}{dataframe or matrix containing the coordinates.}
}
\value{
Returns Quadrant(s) of vector(s)
}
\description{
Find the quadrant(s) of vector(s).
If the vector(s) is a principle axis, it returns the angle(s) in degrees.
}
\examples{
data.frame(x = c(1,-1),y = c(1,-1))|>math_quadrant()
math_quadrant(c(1,0))
math_quadrant(c(1,0,12))
m <- expand.grid(x = -1:1, y = -1:1)
cbind(m, result = math_quadrant(m))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
