% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_quadrant_vector}
\alias{math_quadrant_vector}
\title{Find quadrant of vector}
\usage{
math_quadrant_vector(x)
}
\arguments{
\item{x}{vector containing the coordinates}
}
\value{
Returns Quadrant of vector(s)
}
\description{
find the
}
\details{
Finds the quadrant of vector. If the vector lies on  a principle axis, it returns the angle in degrees.
}
\examples{
math_quadrant_vector(c(1,1))
x <- expand.grid(x = -1:1, y = -1:1)
cbind(x, result = x|>apply(1,math_quadrant_vector))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
