% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_quadrant_vector}
\alias{math_quadrant_vector}
\title{Find quadrant of single vector}
\usage{
math_quadrant_vector(x)
}
\arguments{
\item{x}{vector containing the coordinates. e.g.first x, second y.}
}
\value{
Returns Quadrant of vector(s)
}
\description{
find the
}
\details{
Find the quadrant of vecotr. If it is on principle it returns the axis First: 5, Second: 6
}
\examples{
math_quadrant(c(1,1))
math_quadrant(c(-1,1))
math_quadrant(c(-1,-1))
math_quadrant(c(1,-1))
math_quadrant(c(0,1))
math_quadrant(c(1,0))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
