% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_rot_matrix2d}
\alias{math_rot_matrix2d}
\title{2d rotation matrix}
\usage{
math_rot_matrix2d(angle)
}
\arguments{
\item{angle}{in radiant}
}
\value{
Returns 3D rotation matrix for given axis (unit vector) and angle.
}
\description{
Calculates a 2D rotation matrix for a given \code{angle}. The rotation will be applied by the right hand rule.
}
\details{
\url{https://en.wikipedia.org/wiki/Rotation_matrix}
}
\examples{
math_rot_matrix2d(30/180*pi)
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
