% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_rot_matrix3d}
\alias{math_rot_matrix3d}
\title{3d rotation matrix}
\usage{
math_rot_matrix3d(x, angle)
}
\arguments{
\item{x}{unit vector c(x1,x2,x3) defining the rotation axis}

\item{angle}{angel in radiant}
}
\value{
Returns 3D rotation matrix for given axis (unit vector) and angle.
}
\description{
Calculates a 3D rotation matrix for a given rotation axis \code{x}(unitvector) and an angle  \code{angle}. The rotation will be appied by the right hand rule.
}
\details{
\url{https://en.wikipedia.org/wiki/Rotation_matrix}
\url{https://de.wikipedia.org/wiki/Drehmatrix#Drehmatrizen_des_Raumes_\%E2\%84\%9D\%C2\%B3}
}
\examples{
math_rot_matrix3d(matrix(c(0,0,1)),0.5)
matrix(c(1,2,3))|>
  math_unit_vector()|>
  math_rot_matrix3d(-pi)
math_rot_matrix3d(matrix(c(0,0,1)),0.5)|>
  det()
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
