% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_rot_matrix3d}
\alias{math_rot_matrix3d}
\title{2d rotation matrix}
\usage{
math_rot_matrix3d(x, angle)
}
\arguments{
\item{x}{unit vector c(x1,x2,x3)}

\item{angle}{vector of angel in rad to rotate}
}
\value{
Returns matrix for given angle
}
\description{
calculates a 3D rotation matrix from axis and angle. The axis is defined by \code{x}. The Rotation will be done around the axis defined by
\code {x} and the root. The angle will be appied by the right hand rule.
}
\details{
\url{https://en.wikipedia.org/wiki/Rotation_matrix}
}
\examples{
math_rot_matrix3d(matrix(c(1,2,3)),c(-pi))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
