% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_slerp}
\alias{math_slerp}
\title{slerp  by 3 points and a given radius.}
\usage{
math_slerp(R, x1, x2, cp, nb_points = 10)
}
\arguments{
\item{R}{radius}

\item{x1}{first point vector}

\item{x2}{second point vector}

\item{cp}{position vector, common centrer}

\item{nb_points}{count of points to generate for the radius interpolation}
}
\value{
Returns point matrix with the calculated coordinates \code{x}
}
\description{
Radius interpolation by 3 points and a given radius.
The centrer point is used to for linear function through x1 and cp as well as for linear function through x2 and cp.
The common point therefore is cp and will be used as position vector.
}
\examples{
Testing
p <- t(as.matrix(data.frame(x1 = c(-10,-5,50),x2 = c(-20,10,2),cp = c(10,-10,10))))

m <- math_slerp(R  =35,
                x1 = p["x1",],
                x2 = p["x2",],
                cp = p["cp",],
                nb_points = 20)
m <- rbind(p,m)
#Plot 3D
library(rgl)
plot3d( m[,1], m[,2], m[,3], type = "p", lwd = 2, top = TRUE,
        #col = rainbow(nrow(m)),
        aspect = c(diff(c(min(m[,1]),max(m[,1]))),
                   diff(c(min(m[,2]),max(m[,2]))),
                   diff(c(min(m[,3]),max(m[,3])))
                   ))
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
