% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_sph2cart}
\alias{math_sph2cart}
\title{Spherical to cartesian coordinates}
\usage{
math_sph2cart(tpr)
}
\arguments{
\item{tpr}{c(\eqn{tilt = \theta}, \eqn{pan = \varphi}, radius = r) as vector or matrix}
}
\value{
named vector or matrix
}
\description{
Transform spherical to cartesian coordinates according to international physics convention:
\eqn{\theta} in range 0...pi (0...180 Deg) and \eqn{\varphi} in range 0...2*pi (0...360Deg)
}
\details{
\url{https://de.wikipedia.org/wiki/Kugelkoordinaten#Umrechnungen}
}
\examples{
math_sph2cart(c(1,0.5,0.1))
rbind(c(1,0.5,0.1),
      c(1,0.2,0.6))|>
      math_sph2cart()
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
