% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math_unit_vector}
\alias{math_unit_vector}
\title{Calculate unit vector}
\usage{
math_unit_vector(x)
}
\arguments{
\item{x}{vector or matrix of coordinates}
}
\value{
unit vector(s)
}
\description{
calculate unite vector from \code{x}
}
\examples{
math_unit_vector(c(x = 1,y = 1,z = 1))
math_unit_vector(c(1,2,3))
math_unit_vector(c(0.5,1))

}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
