% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{r_colourise}
\alias{r_colourise}
\title{Colourise text for display in the terminal.}
\usage{
r_colourise(text, fg = "black", bg = NULL)
}
\arguments{
\item{text}{character vector}

\item{fg}{foreground colour, defaults to white}

\item{bg}{background colour, defaults to transparent}
}
\description{
If R is not currently running in a system that supports terminal colours
the text will be returned unchanged.
}
\details{
Allowed colours are: black, blue, brown, cyan, dark gray, green, light
blue, light cyan, light gray, light green, light purple, light red,
purple, red, white, yellow
}
\examples{
# copy paste the code to the terminal
print(r_colourise("text", "red"))
r_colourise("text", "red")|>writeLines()
}
