% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{r_path}
\alias{r_path}
\title{Convert Windows file path to R compatible file path}
\usage{
r_path(path = "clipboard")
}
\arguments{
\item{empty}{no argument needed}
}
\value{
Character vector of R compatible file path
}
\description{
Copy the file path from windows explorer to your clipboard and call the function with no argument.The R compatible file Path will be returned
}
\examples{
Importend: Copy file path first to the cliboard before calling the function.
r_path()
}
