% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{r_path}
\alias{r_path}
\title{Convert Windows file path to R compatible file path}
\usage{
r_path(path = "clipboard")
}
\arguments{
\item{path}{Path can be copied to clipboard}
}
\value{
Character vector of R compatible file path
}
\description{
Copy the file path from windows explorer to your clipboard and call the function with no argument.
The R compatible file Path will be returned
}
\examples{
# Copy file path to the cliboard and call the function.
utils::writeClipboard("C:\\\\Windows")
r_path()
r_path("C:\\\\Windows")
}
