% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{r_tbl_contents}
\alias{r_tbl_contents}
\title{Automatically creates a table of contents for a .Rmd file}
\usage{
r_tbl_contents(FileName)
}
\arguments{
\item{FileName}{file name or connection}
}
\value{
.Rmd file string
}
\description{
Automatically creates a table of contents for a .Rmd file
}
\details{
Scans documents for headings and creates a table of contents (hyper linked). The returned string can directly be written as .Rmd file.
}
