% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmarkdown_utiilities.R
\name{r_tbl_contents}
\alias{r_tbl_contents}
\title{Automatically creates a table of contents for a .Rmd file}
\usage{
r_tbl_contents(c_Rmd, TOC_nb = FALSE)
}
\arguments{
\item{c_Rmd}{Rmd file string}

\item{TOC_nb}{boolean to enable numbering for each heading}
}
\value{
.Rmd file string
}
\description{
Scans documents for headings and creates a table of contents (hyper linked). The returned string can directly be written as .Rmd file.
}
\details{
The function argument is a string of a R markdown .Rmd file which can be loaded via \code{readLines("fileName.Rmd")}
}
\examples{
print(tbl_of_contents.Rmd)
c_rmd <- r_tbl_contents(tbl_of_contents.Rmd)
c_rmd|>write("test.Rmd")
rmarkdown::render("test.Rmd", c("html_document"))
browseURL("test.html")
c_rmd <- r_tbl_contents(tbl_of_contents.Rmd, T)
c_rmd|>write("test.Rmd")
rmarkdown::render("test.Rmd", c("html_document"))
browseURL("test.html")
}
