% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmarkdown_utiilities.R
\name{r_toc_for_Rmd}
\alias{r_toc_for_Rmd}
\title{Automatically creates a table of contents for a .Rmd file string}
\usage{
r_toc_for_Rmd(
  c_Rmd,
  toc_heading_string = "Table of Contents",
  create_nb = TRUE,
  create_top_link = TRUE,
  nb_front = TRUE,
  set_first_heading_level = FALSE,
  pagebreak_level = "non"
)
}
\arguments{
\item{c_Rmd}{Rmd file string}

\item{toc_heading_string}{title string for TOC, e.g "Inhaltsverzeichnis", default is "Table of contents".}

\item{create_nb}{boolean to enable numbering for each heading.}

\item{create_top_link}{boolean to create link below each heading to jump back to the table of contents.}

\item{nb_front}{boolean to have the numbering in front of each heading.}

\item{set_first_heading_level}{set first found heading level to heading 1}

\item{pagebreak_level}{Automatically add page breaks before new heading. Set the level to insert page breaks to e.g. c("non","1","2","3","4","5","6")}
}
\value{
.Rmd file string
}
\description{
Scans . Rmd file string for headings (#) and creates a table of contents (hyper linked). The returned string can directly be written as .Rmd file.
All code section will be excluded for the (#) search.
}
\details{
The function argument is a string of a R markdown .Rmd file which can be read via \code{readLines("fileName.Rmd")}.
}
\examples{
print(tbl_of_contents.Rmd)
c_rmd <- r_toc_for_Rmd(tbl_of_contents.Rmd, create_nb = FALSE)
c_rmd
c_rmd <- r_toc_for_Rmd(tbl_of_contents.Rmd,
                        toc_heading_string = "Inhaltsverzeichnis", create_nb = TRUE)
c_rmd
}
