% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{r_win_path}
\alias{r_win_path}
\title{Convert R path to windows compatible path}
\usage{
r_win_path(path = "clipboard")
}
\arguments{
\item{path}{R path character vector}
}
\value{
Windows compatible character vector
}
\description{
Function to convert R path to windows compatible path.
}
\examples{
r_win_path(tempdir())
}
