% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singal.R
\name{signal_center}
\alias{signal_center}
\title{Center Signal around zero (AC cupling)}
\usage{
signal_center(x)
}
\arguments{
\item{x}{vector}
}
\value{
centred vector \code{x}
}
\description{
centre data point around zero, that the range(\code{x}) is +/- the same value.
}
\examples{
x <- rnorm(100,mean = -0.5)
plot(x)
signal_center(x)|>plot()
}
\author{
Florian Wagner
\email{florian.wagner@wagnius.ch}
}
